requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.24
python-multipart>=0.0.9
streamlit>=1.30
requests>=2.31
python-dotenv>=1.0

sqlalchemy>=2.0
psycopg2-binary>=2.9

pypdf>=4.0

langchain-core>=0.2
langchain-community>=0.2
langchain-openai>=0.1
langchain-postgres>=0.0.12
langchain-huggingface>=0.0.3
langchain-experimental>=0.0.60

sentence-transformers>=3.0
flashrank>=0.2